+++ pi3d/Shape.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import ctypes
+from ctypes import c_float, c_int, c_short
 
 from numpy import array, dot, savez, load
 from math import radians, pi, sin, cos
@@ -36,7 +37,7 @@ class Shape(Loadable):
     self.name = name
     light = light or Light.instance()
     # uniform variables all in one array (for Shape and one for Buffer)
-    self.unif = (ctypes.c_float * 60)(
+    self.unif = (c_float * 60)(
       x, y, z, rx, ry, rz,
       sx, sy, sz, cx, cy, cz,
       0.5, 0.5, 0.5, 5000.0, 0.8, 1.0,
@@ -96,38 +97,39 @@ class Shape(Loadable):
     self.tr1 = array([[1.0, 0.0, 0.0, 0.0],
                       [0.0, 1.0, 0.0, 0.0],
                       [0.0, 0.0, 1.0, 0.0],
-                      [self.unif[0] - self.unif[9], self.unif[1] - self.unif[10], self.unif[2] - self.unif[11], 1.0]])
+                      [self.unif[0] - self.unif[9], self.unif[1] - self.unif[10], self.unif[2] - self.unif[11], 1.0]],
+                      dtype=c_float)
     """translate to position - offset"""
     s, c = sin(radians(self.unif[3])), cos(radians(self.unif[3]))
     self.rox = array([[1.0, 0.0, 0.0, 0.0],
                       [0.0, c, s, 0.0],
                       [0.0, -s, c, 0.0],
-                      [0.0, 0.0, 0.0, 1.0]])
+                      [0.0, 0.0, 0.0, 1.0]], dtype=c_float)
     """rotate about x axis"""
     s, c = sin(radians(self.unif[4])), cos(radians(self.unif[4]))
     self.roy = array([[c, 0.0, -s, 0.0],
                       [0.0, 1.0, 0.0, 0.0],
                       [s, 0.0, c, 0.0],
-                      [0.0, 0.0, 0.0, 1.0]])
+                      [0.0, 0.0, 0.0, 1.0]], dtype=c_float)
     """rotate about y axis"""
     s, c = sin(radians(self.unif[5])), cos(radians(self.unif[5]))
     self.roz = array([[c, s, 0.0, 0.0],
                       [-s, c, 0.0, 0.0],
                       [0.0, 0.0, 1.0, 0.0],
-                      [0.0, 0.0, 0.0, 1.0]])
+                      [0.0, 0.0, 0.0, 1.0]], dtype=c_float)
     """rotate about z axis"""
     self.scl = array([[self.unif[6], 0.0, 0.0, 0.0],
                       [0.0, self.unif[7], 0.0, 0.0],
                       [0.0, 0.0, self.unif[8], 0.0],
-                      [0.0, 0.0, 0.0, 1.0]])
+                      [0.0, 0.0, 0.0, 1.0]], dtype=c_float)
     """scale"""
     self.tr2 = array([[1.0, 0.0, 0.0, 0.0],
                       [0.0, 1.0, 0.0, 0.0],
                       [0.0, 0.0, 1.0, 0.0],
-                      [self.unif[9], self.unif[10], self.unif[11], 1.0]])
+                      [self.unif[9], self.unif[10], self.unif[11], 1.0]], dtype=c_float)
     """translate to offset"""
     self.MFlg = True
-    self.M = (ctypes.c_float * 32)(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
+    self.M = (c_float * 32)(0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
                             0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
@@ -702,7 +704,7 @@ class Shape(Loadable):
   
   def __setstate__(self, state):
     unif_tuple = tuple(state['unif'])
-    self.unif = (ctypes.c_float * 60)(*unif_tuple)
+    self.unif = (c_float * 60)(*unif_tuple)
     self.childModel = state['childModel']
     self.name = state['name']
     self.children = state['children']
